+++ packages/flights-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @voyant-travel/flights-react
 
+## 0.304.0
+
+### Patch Changes
+
+- Updated dependencies [72c2616]
+  - @voyant-travel/finance@0.260.0
+  - @voyant-travel/finance-react@0.304.0
+  - @voyant-travel/relationships-react@0.304.0
+
 ## 0.303.0
 
 ### Patch Changes