+++ packages/legal-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @voyant-travel/legal-react
 
+## 0.304.0
+
+### Patch Changes
+
+- Updated dependencies [57804ad]
+  - @voyant-travel/bookings-react@0.304.0
+  - @voyant-travel/inventory-react@0.186.0
+  - @voyant-travel/distribution-react@0.294.0
+  - @voyant-travel/commerce-react@0.186.0
+  - @voyant-travel/relationships-react@0.304.0
+
 ## 0.303.0
 
 ### Patch Changes