+++ packages/operator-settings-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @voyant-travel/operator-settings-react
 
+## 0.162.0
+
+### Patch Changes
+
+- Updated dependencies [72c2616]
+  - @voyant-travel/finance@0.260.0
+  - @voyant-travel/finance-react@0.304.0
+  - @voyant-travel/admin-app@0.197.0
+
 ## 0.161.0
 
 ### Patch Changes