+++ packages/operator-settings/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @voyant-travel/operator-settings
 
+## 0.18.13
+
+### Patch Changes
+
+- Updated dependencies [72c2616]
+  - @voyant-travel/finance@0.260.0
+  - @voyant-travel/commerce@0.55.1
+
 ## 0.18.12
 
 ### Patch Changes