+++ packages/trips/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @voyant-travel/trips
 
+## 0.237.4
+
+### Patch Changes
+
+- Updated dependencies [72c2616]
+  - @voyant-travel/finance@0.260.0
+  - @voyant-travel/catalog@0.260.1
+  - @voyant-travel/commerce@0.55.1
+  - @voyant-travel/flights@0.240.2
+  - @voyant-travel/inventory@0.42.22
+  - @voyant-travel/operator-settings@0.18.13
+  - @voyant-travel/public-api@0.263.1
+
 ## 0.237.3
 
 ### Patch Changes